+++ src/Recyclarr.TrashLib/Config/Parsing/PostProcessing/IConfigPostProcessor.cs
@@ -1,0 +1,6 @@
+namespace Recyclarr.TrashLib.Config.Parsing.PostProcessing;
+
+public interface IConfigPostProcessor
+{
+    RootConfigYaml Process(RootConfigYaml config);
+}