+++ src/Recyclarr.TrashLib/Config/Secrets/ISecretsProvider.cs
@@ -1,8 +1,6 @@
-using System.Collections.Immutable;
-
 namespace Recyclarr.TrashLib.Config.Secrets;
 
 public interface ISecretsProvider
 {
-    IImmutableDictionary<string, string> Secrets { get; }
+    IReadOnlyDictionary<string, string> Secrets { get; }
 }